Main Java Class
This program uses an iterative process that solves for the final velocity of a block traveling down a ramp with friction. Friction, according to basic Newtonian mechanics, is proportional to normal force. As a result, the frictional force varies with the angle of the block on the ramp. This program splits the travel of the block on the ramp into smaller, assumed linear distances to approximate the final velocity of the block with greater accuracy as the assumed linear distances get smaller.